refactor: 댓글삭제 redux 수정

+++ src/components/detailPage/comments/index.js
@@ -10,25 +10,27 @@ import {
   CommentListWrapper,
 } from "./style";
 
-import DeleteIcon from "@mui/icons-material/Delete";
+import { FiTrash2 } from "react-icons/fi";
 import { getCookie } from "../../../shared/Cookie";
 import CreatedAt from "../../global/createdAt";
 import { useSelector, useDispatch } from "react-redux";
 import { __deleteComment } from "../../../redux/async/asyncComment";
-import { deleteComment } from "../../../redux/modules/commentSlice";
 
 const CommentList = () => {
   const dispatch = useDispatch();
   const comments = useSelector((state) => state.comment.commentList);
-
   const memberId = getCookie("memberId"); //로그인한 유저닉네임
 
   const deleteButton = (commentId) => {
-    const data = {
-      postId: comments[0]["postId"],
-      commentId: commentId,
-    };
-    dispatch(__deleteComment(data));
+    if (window.confirm("정말 삭제하시나요?") === true) {
+      const data = {
+        postId: comments[0]["postId"],
+        commentId: commentId,
+      };
+      dispatch(__deleteComment(data));
+    } else {
+      return;
+    }
   };
 
   return (
@@ -54,15 +56,9 @@ const CommentList = () => {
                     <DeleteBox
                       onClick={() => {
                         deleteButton(comment.commentId);
-                        dispatch(deleteComment(comment.commentId));
                       }}
                     >
-                      <DeleteIcon
-                        fontSize="large"
-                        Width="100%"
-                        height="50%"
-                        display="flex"
-                      />
+                      <FiTrash2 color="red" />
                     </DeleteBox>
                   ) : null}
                 </Side>
